Twitter 0.0.3
[FEATURE][IMPLEMENTATION][CLASS] LoadingFiles class is now more completed and more useful, also ObserverService is now almost completed and ready to work 
there are some bugs that I do not know about them until now because the program was not executed logically yet 
soon there will be a user interface to test methods and see how they work (especially those which related to files)

+++ Twitter/src/Base/TweetingService.java
@@ -5,14 +5,18 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.ObjectOutputStream;
 import java.time.LocalDateTime;
+import java.util.Iterator;
+
+import static Base.LoadingFiles.individualTweets;
 
 public class TweetingService {
     // create a Tweet class
     public void tweeting(Tweet tweet) {
-        String tweetCode = tweet.getId();
+        String tweetCode = tweet.getIdTweet();
         try (FileOutputStream fileOutputStream = new FileOutputStream(tweetCode)) {
             ObjectOutputStream objectOutputStream = new ObjectOutputStream(fileOutputStream);
             objectOutputStream.writeObject(tweet);
+            individualTweets.get(tweet.getSender().getUsername()).add(tweet);
             objectOutputStream.flush();
             objectOutputStream.close();
         } catch (IOException e) {
@@ -21,14 +25,26 @@ public class TweetingService {
     }
     // assign null to object reference11
     public boolean deleting(Tweet deletedTweet) {
-        String tweetCode = deletedTweet.getId() + ".txt";
-        File file = new File(tweetCode);
-            if (file.delete()) {
-                return true;
+        boolean flag = false;
+        if (deletedTweet.getSender().getUsername() != null) {
+            Iterator<Tweet> iterator = individualTweets.get(deletedTweet.getSender().getUsername()).iterator();
+            while (iterator.hasNext()) {
+                Tweet tweetTemp = iterator.next();
+                if (tweetTemp.equals(deletedTweet)) {    // to make sure that tweet is really existed
+                    String tweetCode = deletedTweet.getIdTweet() + ".txt";
+                    File file = new File(tweetCode);
+                    if (file.delete()) {
+                        individualTweets.get(deletedTweet.getSender().getUsername()).remove(deletedTweet);
+                        flag = true;
+                    }
+                }
             }
-            else
-                return false;
+            return flag;
+        }
+        else
+            return false;
     }
+    /*
     // assign tweet to another account
     public void retweeting(Tweet retweetedTweet, Account retweeter) {
         Retweet retweet = new Retweet("Retweeted from " + retweetedTweet.getSender(),
@@ -39,9 +55,10 @@ public class TweetingService {
                 0);
         retweetedTweet.setRetweets(retweetedTweet.getRetweets()+1);
         retweet.setRetweeter(retweeter);
-        retweet.setId(retweetedTweet.getId() + "-retweeted");
+        retweet.setIdTweet(retweetedTweet.getIdTweet() + "-retweeted");
         tweeting(retweetedTweet);
-    }
+    }*/
+
     // add up like number for the tweet
     public void liking(Tweet likedTweet) {
         likedTweet.setLikes(likedTweet.getLikes()+1);
